requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
openai>=1.30.0
anthropic>=0.28.0
python-pptx>=0.6.23
matplotlib>=3.9.0
plotly>=5.22.0
pandas>=2.2.0
numpy>=1.26.0
Pillow>=10.3.0
requests>=2.32.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
httpx>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.1
openpyxl>=3.1.4
PyMuPDF>=1.24.0
rich>=13.7.0
typer>=0.12.3
